requirements_ragas.txt
# RAGAS评估依赖
ragas>=0.1.0
datasets>=2.14.0
evaluate>=0.4.0
sacrebleu>=2.5.0

# 现有项目依赖
openai>=1.0.0
pinecone-client>=2.2.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0

# 可选依赖
umap-learn>=0.5.0
bert-score>=0.3.0
rouge-score>=0.1.0 